Added a target manager
Added a target manager. this holds the information of the currently
loaded target. This is held outside of the target game object. If
another target is loaded during the time that another target is already
loaded, it will unload the target already loaded before then loading
the new target. Confusing right? Hopefully that fixes it.