Hunter Arrow Actions
These scripts are for the hunter. There are some animation events that
trigger gameobjects to be enabled/disabled. This creates the effect of
the hunter grabbing his bow, reaching into his quiver for an arrow,
moving the arrow on to the bow (switching the arrow’s hand), pulling
back, and then an arrow is fired that "fails" as the hunter cannot
reach the deer.