Tsikwa's Objects
Tsikwa hits the deer with several objects and the height the deer jump
reduces by half. An animation event instantiates new objects (such as
Apron, Shawl, Paint Bag). These GameObjects then move towards a
hitpoint on the deer. When the distance between the object and the
hitpoint is a certain amount, the deer jump to a lesser height.