.mp3 files
converted all audiofiles from .wav to .mp3 and compressed (cut down
file size significantly). Updated script so that previously established
naming convention for targets are used in order to retrieve
corresponding .mp3 file, therefore reducing redundancy in the editor
and allowing for cleaner interface when selecting speakers in target
controller. Debug.Log() is called in case there are any errors in
naming. Loose naming conventions are also accounted for since the
certain chars are replaced in case target is named 5-11,12,13 or
5-11.12.13 for example, will now be 5-11_12_13 as per name of .mp3
file. Certain Onboarding fixes, and an unused AudioLoader that may come
in handy later if using AssetBundles.